HTML5 Fractal Playground 1.1
Completely redo the way colors are computed.  In the past, there was a
maximum number of iterations to test for escape and that was one extreme
color, and the other extreme color was for points that escape in only
one iteration.

In the new version, one extreme color is for points that take the
maximum number of iterations for the given plot, *which is now allowed
to vary between renderings*.  The colors are chosen to show the
difference across the range of escapes we see on the plot.